renames domain.py into vm.py
